I create a new view for the main page and I create a new HTML file and URL in urlpatterns for it.
The name of the new view is index_main_page and the only return template and name of this template is index.html which is located under the template folder such as other HTML files. I have used bootstrap 4 CDNs or links in index.html. The main page can be used for navigation to upload, table, and statistics pages with the menu and users can see source code with 'Learn more' button. Also, I create a new URL in urlpatterns which internal name of this URL is index_main_page. This main page is returned to the user when the user only types the name of the webpage.